Change getKeyIntValueForDump to use intptr_t rather than long

+++ stdlib/public/runtime/MetadataCache.h
@@ -217,7 +217,7 @@ template <class ValueTy> class MetadataCache {
       return KeyDataRef::forArguments(getKeyDataBuffer(), KeyLength);
     }
 
-    long getKeyIntValueForDump() const {
+    intptr_t getKeyIntValueForDump() const {
       return Hash;
     }
